Add the ability to delete and restore rows from the UI. (#1011)
Support deleting rows from the UI.
- Individual rows have a delete button to add the special delete label.
- The schema panel now shows number of deleted, with a button to view
the deleted.
- Delete in filter now happens.
- Export auto-adds the deleted label.

https://huggingface.co/spaces/lilacai/nikhil_staging


https://github.com/lilacai/lilac/assets/1100749/35f43853-6875-4e59-afe6-dca9aaaadada

+++ lilac/data/dataset_duckdb.py
@@ -2163,9 +2163,10 @@ class DatasetDuckDB(Dataset):
           [(x,) for x in remove_row_ids],
         )
         conn.commit()
-        count = conn.execute(f'SELECT COUNT(*) FROM "{name}"').fetchone()[0]
-        if count == 0:
-          delete_file(labels_filepath)
+        if name != DELETED_LABEL_NAME:
+          count = conn.execute(f'SELECT COUNT(*) FROM "{name}"').fetchone()[0]
+          if count == 0:
+            delete_file(labels_filepath)
 
     if remove_row_ids and name == DELETED_LABEL_NAME:
       self.stats.cache_clear()
